+++ src/manifolds/FixedRankMatrices.jl
@@ -1,5 +1,5 @@
 @doc raw"""
-    FixedRankMatrices{T,𝔽} <: AbstractDecoratorManifold{𝔽}
+    FixedRankMatrices{𝔽, T} <: AbstractDecoratorManifold{𝔽}
 
 The manifold of ``m×n`` real-valued or complex-valued matrices of fixed rank ``k``, i.e.
 ````math
@@ -36,7 +36,7 @@ on ``ℝ^{m×n}`` to the tangent bundle [Vandereycken:2013](@cite).
 
 Generate the manifold of `m`-by-`n` (`field`-valued) matrices of rank `k`.
 """
-struct FixedRankMatrices{T, 𝔽} <: AbstractDecoratorManifold{𝔽}
+struct FixedRankMatrices{𝔽, T} <: AbstractDecoratorManifold{𝔽}
     size::T
 end
 
@@ -48,7 +48,7 @@ function FixedRankMatrices(
         parameter::Symbol = :type,
     )
     size = wrap_type_parameter(parameter, (m, n, k))
-    return FixedRankMatrices{typeof(size), field}(size)
+    return FixedRankMatrices{field, typeof(size)}(size)
 end
 
 @doc raw"""
@@ -444,10 +444,10 @@ function embed!(::FixedRankMatrices, Y, p::SVDMPoint, X::UMVTangentVector)
     return mul!(Y, p.U, X.Vt, true, true)
 end
 
-function get_embedding(::FixedRankMatrices{TypeParameter{Tuple{m, n, k}}, 𝔽}) where {m, n, k, 𝔽}
+function get_embedding(::FixedRankMatrices{𝔽, TypeParameter{Tuple{m, n, k}}}) where {m, n, k, 𝔽}
     return Euclidean(m, n; field = 𝔽)
 end
-function get_embedding(M::FixedRankMatrices{Tuple{Int, Int, Int}, 𝔽}) where {𝔽}
+function get_embedding(M::FixedRankMatrices{𝔽, Tuple{Int, Int, Int}}) where {𝔽}
     m, n, k = get_parameter(M.size)
     return Euclidean(m, n; field = 𝔽, parameter = :field)
 end
@@ -544,7 +544,7 @@ of dimension `m`x`n` of rank `k`, namely
 
 where ``\dim_ℝ 𝔽`` is the [`real_dimension`](@extref `ManifoldsBase.real_dimension-Tuple{ManifoldsBase.AbstractNumbers}`) of `𝔽`.
 """
-function manifold_dimension(M::FixedRankMatrices{<:Any, 𝔽}) where {𝔽}
+function manifold_dimension(M::FixedRankMatrices{𝔽}) where {𝔽}
     m, n, k = get_parameter(M.size)
     return (m + n - k) * k * real_dimension(𝔽)
 end
@@ -778,11 +778,11 @@ end
 
 function Base.show(
         io::IO,
-        ::FixedRankMatrices{TypeParameter{Tuple{m, n, k}}, 𝔽},
+        ::FixedRankMatrices{𝔽, TypeParameter{Tuple{m, n, k}}},
     ) where {m, n, k, 𝔽}
     return print(io, "FixedRankMatrices($(m), $(n), $(k), $(𝔽))")
 end
-function Base.show(io::IO, M::FixedRankMatrices{Tuple{Int, Int, Int}, 𝔽}) where {𝔽}
+function Base.show(io::IO, M::FixedRankMatrices{𝔽, Tuple{Int, Int, Int}}) where {𝔽}
     m, n, k = get_parameter(M.size)
     return print(io, "FixedRankMatrices($(m), $(n), $(k), $(𝔽); parameter=:field)")
 end