+++ src/manifolds/GrassmannStiefel.jl
@@ -28,7 +28,7 @@ end
 ManifoldsBase.@manifold_element_forwards StiefelPoint value
 ManifoldsBase.@manifold_vector_forwards StiefelTangentVector value
 ManifoldsBase.@default_manifold_fallbacks Stiefel StiefelPoint StiefelTangentVector value value
-ManifoldsBase.@default_manifold_fallbacks (Stiefel{<:Any, ℝ}) StiefelPoint StiefelTangentVector value value
+ManifoldsBase.@default_manifold_fallbacks (Stiefel{ℝ}) StiefelPoint StiefelTangentVector value value
 ManifoldsBase.@default_manifold_fallbacks Grassmann StiefelPoint StiefelTangentVector value value
 
 function ManifoldsBase.get_forwarding_type(::Grassmann, ::typeof(isapprox), p)
@@ -98,10 +98,10 @@ function exp!(M::Grassmann, q, p, X)
     return copyto!(q, Array(qr(z).Q))
 end
 
-function get_embedding(::Grassmann{TypeParameter{Tuple{n, k}}, 𝔽}) where {n, k, 𝔽}
+function get_embedding(::Grassmann{𝔽, TypeParameter{Tuple{n, k}}}) where {n, k, 𝔽}
     return Stiefel(n, k, 𝔽)
 end
-function get_embedding(M::Grassmann{Tuple{Int, Int}, 𝔽}) where {𝔽}
+function get_embedding(M::Grassmann{𝔽, Tuple{Int, Int}}) where {𝔽}
     n, k = get_parameter(M.size)
     return Stiefel(n, k, 𝔽; parameter = :field)
 end
@@ -296,7 +296,7 @@ rand(M::Grassmann; σ::Real = 1.0)
 
 function Random.rand!(
         rng::AbstractRNG,
-        M::Grassmann{<:Any, 𝔽},
+        M::Grassmann{𝔽},
         pX;
         σ::Real = one(real(eltype(pX))),
         vector_at = nothing,
@@ -404,9 +404,9 @@ The formula reads [Rentmeesters:2011](@cite)
 R(X,Y)Z = (XY^\mathrm{T} - YX^\mathrm{T})Z + Z(Y^\mathrm{T}X - X^\mathrm{T}Y).
 ```
 """
-riemann_tensor(::Grassmann{<:Any, ℝ}, p, X, Y, Z)
+riemann_tensor(::Grassmann{ℝ}, p, X, Y, Z)
 
-function riemann_tensor!(::Grassmann{<:Any, ℝ}, Xresult, p, X, Y, Z)
+function riemann_tensor!(::Grassmann{ℝ}, Xresult, p, X, Y, Z)
     XYᵀ = X * Y'
     YXᵀ = XYᵀ'
     YᵀX = Y' * X
@@ -415,10 +415,10 @@ function riemann_tensor!(::Grassmann{<:Any, ℝ}, Xresult, p, X, Y, Z)
     return Xresult
 end
 
-function Base.show(io::IO, ::Grassmann{TypeParameter{Tuple{n, k}}, 𝔽}) where {n, k, 𝔽}
+function Base.show(io::IO, ::Grassmann{𝔽, TypeParameter{Tuple{n, k}}}) where {n, k, 𝔽}
     return print(io, "Grassmann($(n), $(k), $(𝔽))")
 end
-function Base.show(io::IO, M::Grassmann{Tuple{Int, Int}, 𝔽}) where {𝔽}
+function Base.show(io::IO, M::Grassmann{𝔽, Tuple{Int, Int}}) where {𝔽}
     n, k = get_parameter(M.size)
     return print(io, "Grassmann($(n), $(k), $(𝔽); parameter=:field)")
 end